Fixed I/O for strings and minor changes in the parser
Previously, the input( ) and print( ) statements would not work, as the compiler would fail to add the reference address to scanf( ) and printf( ) respectively. Now that has been fixed and strings can be input and printed out. Made some minor changes in simc_parser.py as well. Changed the string formatting from .format, to % operator. Added rawstring character before the multiline string comments to prevent warnings that involved the backslash character. Added and OR condition to line 650 which dealt with strings and character arrays.